+++ CHANGELOG.md
@@ -38,6 +38,7 @@
 * Add About dialog
 * Fix some mod info text being selectable
 * Remove MTS version from window title
+* Convert profile editing UI into separate window, renamed to Mod Lists
 
 #### v3.30.3 ####
 * Fix package information being lost for patched classes (Alchyr)